Add TOR in markdown format
I've converted the TOR to markdown.

I did make a tiny change from the TOR as accepted by the PSP - I removed
the address for the mailing list. It's a closed list but it seemed
unnecessary for the address to be publicly available.

Before being merged, the membership form needs to be added to the site
and linked from this document.